backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
xgboost
joblib
requests
python-dotenv
matplotlib
seaborn
pydantic
google-auth
google-auth-oauthlib
google-auth-httplib2
passlib[bcrypt]
python-multipart
python-jose[cryptography]
aiofiles
